Adding dependent dropdown in home page

+++ transfer/views/remove_data.py
@@ -4,15 +4,20 @@ from ..models.model_transfer_course import TransferCourse
 from ..models.model_major_requirement import Major_requirement
 from ..models.model_transferevaluation import Transferevaluation
 from ..models.model_approver import Approver
+from django.urls import reverse
+from django.http import HttpResponseRedirect
 
-from django.shortcuts import render, get_object_or_404
+from django.shortcuts import render, get_object_or_404, redirect
 
 
 def remove_all_data(request):
-    Major.objects.all().delete()
-    School.objects.all().delete()
-    TransferCourse.objects.all().delete()
-    Major_requirement.objects.all().delete()
-    Transferevaluation.objects.all().delete()
-    Approver.objects.all().delete()
-    return render(request, 'home.html')
+
+    if request.method == "POST":
+        Major.objects.all().delete()
+        School.objects.all().delete()
+        TransferCourse.objects.all().delete()
+        Major_requirement.objects.all().delete()
+        Transferevaluation.objects.all().delete()
+        Approver.objects.all().delete()
+        return HttpResponseRedirect(reverse("home"))
+    return render(request, 'removeall.html')
